+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The change log is available [on GitHub][2].
 1.0.0
 =====
 
-* TODO
+* [#100](https://github.com/serokell/o-clock/issues/100):
+  Add `Hashable` instance for `Time`.
 
 0.1.1
 =====